# openapi.yaml
openapi: 3.0.3
info:
  title: Food Delivery Platform API
  description: |-
    A comprehensive food delivery platform with microservices architecture.
    
    ## Services Overview
    - **Auth Service**: User authentication and authorization
    - **Order Service**: Order management and lifecycle
    - **Payment Service**: Payment processing and wallet management
    - **Restaurant Service**: Menu management and restaurant operations
    - **Delivery Service**: Delivery assignment and tracking
    - **Tracking Service**: Real-time location tracking
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```
  version: 1.0.0
  contact:
    name: Food Delivery Platform Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Kong API Gateway (Development)
  - url: https://api.fooddelivery.com
    description: Production API Gateway
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
        '401':
          description: Invalid credentials
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: password123
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request or user already exists
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate the current session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized

  # Order Management Endpoints
  /api/v1/orders:
    post:
      tags:
        - Orders
      summary: Place a new order
      description: Create a new food order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - restaurant_id
                - items
              properties:
                user_id:
                  type: string
                  example: user123
                restaurant_id:
                  type: string
                  example: rest456
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/MenuItem'
                  minItems: 1
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id:
                    type: string
                    example: order_1640995200000000000
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded

    get:
      tags:
        - Orders
      summary: Get user orders
      description: Retrieve all orders for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
  /api/v1/orders/{orderId}/status:
    put:
      tags:
        - Orders
      summary: Update order status
      description: Update the status of an existing order (Admin/Restaurant only)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
          example: order_1640995200000000000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [confirmed, preparing, ready, picked_up, delivered, cancelled]
                  example: confirmed
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Order not found

  # Payment Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: user123
        email:
          type: string
          format: email
          example: user@example.com
        role:
          type: string
          enum: [user, admin, restaurant, delivery_agent]
          example: user
    Order:
      type: object
      properties:
        id:
          type: string
          example: order_1640995200000000000
        user_id:
          type: string
          example: user123
        restaurant_id:
          type: string
          example: rest456
        status:
          type: string
          enum: [created, confirmed, preparing, ready, picked_up, delivered, cancelled]
          example: confirmed
        amount:
          type: number
          format: float
          example: 25.99
        created_at:
          type: string
          format: date-time
          example: "2023-12-01T10:00:00Z"
    MenuItem:
      type: object
      properties:
        id:
          type: string
          example: item_1640995200000000000
        name:
          type: string
          example: "Margherita Pizza"
        price:
          type: number
          format: float
          example: 12.99
        available:
          type: boolean
          example: true
...